feat(talaria-proc-lib): change pid paramter to i64 from u32 If Hermes is deployed on a 32 bit operating machine, having the `pid` converted to usize may truncate the `pid` to value that was not originally passed to the kill() function. This ensures that the `pid` is preserved in the error message to the user.

+++ talaria/src/stdlib/error.rs
@@ -119,7 +119,7 @@ pub mod error {
         ProcFailedToStartProcess { command: String },
 
         #[strum(props(class = "proc", name = "BadPid"), to_string = "bad pid: {pid}")]
-        ProcBadPid { pid: usize },
+        ProcBadPid { pid: i64 },
     }
 
     impl<T> Into<Result<T, Box<EvalAltResult>>> for ScriptError {
